AF-1656: Binding Column 1 Properties
