+++ dashbuilder/dashbuilder-client/dashbuilder-renderers/dashbuilder-renderer-c3/src/main/java/org/dashbuilder/renderer/c3/client/charts/bubble/C3BubbleChartDisplayer.java
@@ -19,6 +19,8 @@ import org.dashbuilder.displayer.DisplayerAttributeDef;
 import org.dashbuilder.displayer.DisplayerAttributeGroupDef;
 import org.dashbuilder.displayer.DisplayerConstraints;
 import org.dashbuilder.renderer.c3.client.C3Displayer;
+import org.dashbuilder.renderer.c3.client.C3XYDisplayer;
+import org.dashbuilder.renderer.c3.client.charts.area.C3AreaChartDisplayer;
 import org.dashbuilder.renderer.c3.client.jsbinding.C3DataInfo;
 import org.dashbuilder.renderer.c3.client.jsbinding.C3Point;
 
@@ -29,7 +31,7 @@ import jsinterop.base.Any;
 import jsinterop.base.Js;
 
 @Dependent
-public class C3BubbleChartDisplayer extends C3Displayer<C3BubbleChartDisplayer.View> {
+public class C3BubbleChartDisplayer extends C3XYDisplayer<C3AreaChartDisplayer.View> {
     
     public interface View extends C3Displayer.View<C3BubbleChartDisplayer> {
     }
