+++ dashbuilder/dashbuilder-client/dashbuilder-renderers/dashbuilder-renderer-c3/src/main/java/org/dashbuilder/renderer/c3/client/charts/line/C3LineChartDisplayer.java
@@ -10,9 +10,11 @@ import org.dashbuilder.displayer.DisplayerAttributeDef;
 import org.dashbuilder.displayer.DisplayerAttributeGroupDef;
 import org.dashbuilder.displayer.DisplayerConstraints;
 import org.dashbuilder.renderer.c3.client.C3Displayer;
+import org.dashbuilder.renderer.c3.client.C3XYDisplayer;
+import org.dashbuilder.renderer.c3.client.charts.area.C3AreaChartDisplayer;
 
 @Dependent
-public class C3LineChartDisplayer extends C3Displayer<C3LineChartDisplayer.View> {
+public class C3LineChartDisplayer extends C3XYDisplayer<C3AreaChartDisplayer.View> {
     
     public interface View extends C3Displayer.View<C3LineChartDisplayer> {
