+++ dashbuilder/dashbuilder-client/dashbuilder-renderers/dashbuilder-renderer-c3/src/main/java/org/dashbuilder/renderer/c3/client/jsbinding/C3ChartConf.java
@@ -47,6 +47,9 @@ public class C3ChartConf {
     public native void setAxis(C3AxisInfo axis);
     
     @JsProperty
+    public native C3AxisInfo getAxis();
+    
+    @JsProperty
     public native void setGrid(C3Grid grid);
     
     @JsProperty
